Update StringTableEsTreatments.xml
Corrected some issues, as well as continuing the translation until line 202. The file is not completly translated yet, there are some words that may need to be changed (E.g.: Ward not translated yet). Some of the Required Rooms have been translated for the entire file since they are the same on every line.